# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Regulatory Compliance AI",
    "description": "Seven agents turn dense regulation into testable obligations \u2014 each with a citation verified in CODE against the source text \u2014 score their impact, and assemble audit-ready evidence packages. A human always signs off; this API determines nothing.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/domains/{domain}/regulations": {
      "get": {
        "tags": [
          "domains"
        ],
        "summary": "List Regulations",
        "operationId": "list_regulations_api_domains__domain__regulations_get",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegulationSummary"
                  },
                  "title": "Response List Regulations Api Domains  Domain  Regulations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/domains/{domain}/audit-types": {
      "get": {
        "tags": [
          "domains"
        ],
        "summary": "List Audit Types",
        "description": "Audit type -> the obligations that audit tests. Drives the Audit Prep screen.",
        "operationId": "list_audit_types_api_domains__domain__audit_types_get",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "title": "Response List Audit Types Api Domains  Domain  Audit Types Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/domains/{domain}/cases": {
      "get": {
        "tags": [
          "domains"
        ],
        "summary": "List Cases",
        "operationId": "list_cases_api_domains__domain__cases_get",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Cases Api Domains  Domain  Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/monitor/run": {
      "post": {
        "tags": [
          "WF-01 Regulatory Monitor"
        ],
        "summary": "Run Monitor",
        "description": "Fetch \u2192 Classify \u2192 Extract (+VERIFY CITATION) \u2192 Map to departments \u2192 Alert.\n\n\u26a0\ufe0f COST + LATENCY: this is `2N + 1` LLM calls for N regulations. For the full Energy pack that\nis 23 calls and takes minutes. **In production this MUST be a background job** with polling or\nSSE \u2014 a blocking POST will time out behind any real gateway. It is synchronous here so the\ncontract is legible; do not ship it this way.",
        "operationId": "run_monitor_api_monitor_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/impact/run": {
      "post": {
        "tags": [
          "WF-02 Obligation Impact"
        ],
        "summary": "Run Impact",
        "description": "Decompose \u2192 Cross-reference \u2192 Score (cost/ops/timeline/penalty) \u2192 Executive report.\n\nExactly 4 LLM calls. Predictable and cheap.",
        "operationId": "run_impact_api_impact_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/run": {
      "post": {
        "tags": [
          "WF-03 Audit Prep (FLAGSHIP)"
        ],
        "summary": "Run Audit",
        "description": "Supervisor plans \u2192 Evidence Collector \u2192 Gap Analyzer \u2192 Response Drafter \u2192 Supervisor reviews.\n\nProduces the traceable chain a regulator asks for when it says \"show me\":\n\n    regulation \u2192 obligation \u2192 evidence doc \u2192 gap \u2192 owner + date \u2192 drafted response\n\nterminating in a 0\u2013100 readiness score. 5 LLM calls.",
        "operationId": "run_audit_api_audit_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/analyze": {
      "post": {
        "tags": [
          "WF-04 Case Analytics"
        ],
        "summary": "Analyze Cases",
        "description": "RAG over the enforcement corpus. 1 LLM call.",
        "operationId": "analyze_cases_api_cases_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Analyze Cases Api Cases Analyze Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/stats": {
      "get": {
        "tags": [
          "WF-04 Case Analytics"
        ],
        "summary": "Case Stats",
        "description": "Aggregate stats + penalty timeline. NO LLM CALL \u2014 the dashboard renders free.",
        "operationId": "case_stats_api_cases_stats_get",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "energy_utilities",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Case Stats Api Cases Stats Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImpactRequest": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain",
            "default": "energy_utilities"
          },
          "regulation_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regulation Index",
            "description": "Index into GET /api/domains/{domain}/regulations"
          },
          "regulation_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regulation Text",
            "description": "Or paste raw regulatory text"
          },
          "regulation_source": {
            "type": "string",
            "title": "Regulation Source",
            "default": "Unknown"
          }
        },
        "type": "object",
        "title": "ImpactRequest"
      },
      "ImpactResponse": {
        "properties": {
          "review_status": {
            "type": "string",
            "const": "PENDING_HUMAN_REVIEW",
            "title": "Review Status",
            "default": "PENDING_HUMAN_REVIEW"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "AI-generated decision support. Not a compliance determination. Every obligation, gap and drafted response requires review and sign-off by qualified personnel before use, and no content may be submitted to a regulator without that sign-off."
          },
          "atomic_obligations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Atomic Obligations"
          },
          "cross_references": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Cross References"
          },
          "impact_scores": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Impact Scores"
          },
          "report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Report",
            "description": "executive_summary, cost range, earliest_deadline, key_risks, recommended_actions[], board_attention_items[], regulatory_strategy"
          }
        },
        "type": "object",
        "required": [
          "atomic_obligations",
          "cross_references",
          "impact_scores",
          "report"
        ],
        "title": "ImpactResponse"
      },
      "MonitorRequest": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain",
            "default": "energy_utilities"
          },
          "source_filter": {
            "type": "string",
            "title": "Source Filter",
            "description": "Regulator code (e.g. 'OEIS', 'CPUC') or 'all'",
            "default": "all"
          }
        },
        "type": "object",
        "title": "MonitorRequest"
      },
      "MonitorResponse": {
        "properties": {
          "review_status": {
            "type": "string",
            "const": "PENDING_HUMAN_REVIEW",
            "title": "Review Status",
            "default": "PENDING_HUMAN_REVIEW"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "AI-generated decision support. Not a compliance determination. Every obligation, gap and drafted response requires review and sign-off by qualified personnel before use, and no content may be submitted to a regulator without that sign-off."
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "type": "array",
            "title": "Alerts"
          },
          "obligations": {
            "items": {
              "$ref": "#/components/schemas/Obligation"
            },
            "type": "array",
            "title": "Obligations"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "regulations_scanned": {
            "type": "integer",
            "title": "Regulations Scanned"
          }
        },
        "type": "object",
        "required": [
          "alerts",
          "obligations",
          "provenance",
          "regulations_scanned"
        ],
        "title": "MonitorResponse"
      },
...
      "RegulationSummary": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "published_date": {
            "type": "string",
            "title": "Published Date"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "index",
          "source",
          "title",
          "published_date",
          "text"
        ],
        "title": "RegulationSummary"
      },
...
    }
  }
}